+++ sassc.py
@@ -7,7 +7,7 @@ This provides SassC_ compliant CLI executable named :program:`sassc`:
 .. sourcecode:: console
 
    $ sassc
-   Usage: sassc [options] SCSS_FILE...
+   Usage: sassc [options] SCSS_FILE [CSS_FILE]
 
 There are options as well:
 
@@ -25,6 +25,14 @@ There are options as well:
 
    Path to find images.  Default is the current directory (:file:`./`).
 
+.. option:: -m, -g, --sourcemap
+
+   Emit source map.  Requires the second argument (output CSS filename).
+   The filename of source map will be the output CSS filename followed by
+   :file:`.map`.
+
+   .. versionadded:: 0.4.0
+
 .. option:: -v, --version
 
    Prints the program version.
@@ -38,6 +46,7 @@ There are options as well:
 """
 from __future__ import print_function
 
+import functools
 import optparse
 import sys
 
@@ -45,14 +54,20 @@ from sass import __version__ as VERSION, OUTPUT_STYLES, CompileError, compile
 
 
 def main(argv=sys.argv, stdout=sys.stdout, stderr=sys.stderr):
-    parser = optparse.OptionParser(usage='%prog [options] SCSS_FILE...',
-                                   version='%prog ' + VERSION)
+    parser = optparse.OptionParser(
+        usage='%prog [options] SCSS_FILE [OUT_CSS_FILE]',
+        version='%prog ' + VERSION
+    )
     output_styles = list(OUTPUT_STYLES)
     output_styles = ', '.join(output_styles[:-1]) + ', or ' + output_styles[-1]
     parser.add_option('-s', '--output-style', metavar='STYLE', type='choice',
                       choices=list(OUTPUT_STYLES), default='nested',
                       help='Coding style of the compiled result.  Choose one '
                            'of ' + output_styles + '. [default: %default]')
+    parser.add_option('-m', '-g', '--sourcemap', dest='source_map',
+                      action='store_true', default=False,
+                      help='Emit source map.  Requires the second argument '
+                           '(output css filename).')
     parser.add_option('-I', '--include-path', metavar='DIR',
                       dest='include_paths', action='append',
                       help='Path to find "@import"ed (S)CSS source files.  '
@@ -60,30 +75,56 @@ def main(argv=sys.argv, stdout=sys.stdout, stderr=sys.stderr):
     parser.add_option('-i', '--image-path', metavar='DIR', default='./',
                       help='Path to find images. [default: %default]')
     options, args = parser.parse_args(argv[1:])
+    error = functools.partial(print,
+                              parser.get_prog_name() + ': error:',
+                              file=stderr)
     if not args:
         parser.print_usage(stderr)
-        print(parser.get_prog_name() + ': error:', 'too few arguments',
-              file=stderr)
+        error('too few arguments')
         return 2
-    elif len(args) > 1:
+    elif len(args) > 2:
         parser.print_usage(stderr)
-        print(parser.get_prog_name() + ': error:', 'too many arguments',
-              file=stderr)
+        error('too many arguments')
         return 2
-    for filename in args:
-        try:
+    filename = args[0]
+    if options.source_map and len(args) < 2:
+        parser.print_usage(stderr)
+        error('-m/-g/--sourcemap requires the second argument, the output '
+              'css filename.')
+        return 2
+    try:
+        if options.source_map:
+            source_map_filename = args[1] + '.map'  # FIXME
+            css, source_map = compile(
+                filename=filename,
+                output_style=options.output_style,
+                source_comments='map',
+                source_map_filename=source_map_filename,
+                include_paths=options.include_paths,
+                image_path=options.image_path
+            )
+        else:
+            source_map_filename = None
+            source_map = None
             css = compile(
                 filename=filename,
                 output_style=options.output_style,
                 include_paths=options.include_paths,
                 image_path=options.image_path
             )
-        except CompileError as e:
-            print(parser.get_prog_name() + ': error:', e,
-                  file=stderr)
-            return 1
-        else:
+    except CompileError as e:
+        print(parser.get_prog_name() + ': error:', e,
+              file=stderr)
+        return 1
+    else:
+        if len(args) < 2:
             print(css, file=stdout)
+        else:
+            with open(args[1], 'w') as f:
+                print(css, file=f)
+        if source_map_filename:
+            with open(source_map_filename, 'w') as f:
+                f.write(source_map)
     return 0